VM, configuration rework, ts application refactor
