+++ .gitignore
@@ -25,3 +25,6 @@ node_modules/
 
 # Misc
 .pre-commit-config.yaml
+
+# Vm
+*.qcow2
