+++ components/cardano-node/hooks/build
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-docker build . --build-context config=../../configurations/ -t $IMAGE_NAME
+docker build . --build-context config=../../cardano-configurations/ -t $IMAGE_NAME
