+++ nix/host/default.nix
@@ -14,6 +14,7 @@ let
           ./disko.nix
           ./configuration.nix
           ./scheduled-tasks.nix
+          ./docker.nix
         ];
         users.users.root.openssh.authorizedKeys.keys = sshAuthorizedKeys;
         _module.args.flake = config.perSystem system;
@@ -23,5 +24,16 @@ in
 {
   flake.nixosConfigurations.cardanow = mkHost {
     sshAuthorizedKeys = import ./ssh-authorized-keys.nix;
+  };
+  perSystem = { lib, ... }: {
+    apps.vm.program =
+      let
+        cardanow-vm = config.flake.nixosConfigurations.cardanow.extendModules {
+          modules = [
+            ./vm.nix
+          ];
+        };
+      in
+      "${lib.getExe cardanow-vm.config.system.build.vm}";
   };
 }
