+++ nix/host/scheduled-tasks.nix
@@ -1,4 +1,4 @@
-{ lib, flake, ... }: {
+{ lib, flake, config, ... }: {
   systemd = {
     timers.cardanow = {
       description = "Run cardanow every 60 minutes";
@@ -13,10 +13,16 @@
     services.cardanow = {
       description = "cardanow";
 
+      path = [ config.virtualisation.docker.package ];
+
       serviceConfig = {
         Type = "oneshot";
-        User = "root"; # TODO is it needed to be executed by root?
+        user = "cardanow";
+        group = "cardanow";
         ExecStart = lib.getExe flake.packages.cardanow;
+        StateDirectory = config.users.users.cardanow.home;
+        WorkingDirectory = config.users.users.cardanow.home;
+        Restart = "on-failure";
       };
     };
   };
